refactor(data): split IndexFactory into separate file and add name field
- Extract IndexFactory from data_processor.py to index_factory.py
- Add name parameter to DataProcessor, DataCollector, DataManager, DataSaver
- Standardized logger naming to f"{self.__class__.__name__}_{self.name}"
